Refactor TabSettings screen
Separate biometric logic in BiometricService file
Remove unused PackageInfo logic
Create a components folder to hold SettingsListTile and EraseWalletDataButton widgets
