+++ lib/di.dart
@@ -4,6 +4,7 @@ import 'package:qubic_wallet/resources/qubic_cmd.dart';
 import 'package:qubic_wallet/resources/qubic_hub.dart';
 import 'package:qubic_wallet/resources/qubic_li.dart';
 import 'package:qubic_wallet/resources/secure_storage.dart';
+import 'package:qubic_wallet/services/biometric_service.dart';
 import 'package:qubic_wallet/stores/application_store.dart';
 import 'package:qubic_wallet/stores/explorer_store.dart';
 import 'package:qubic_wallet/stores/qubic_hub_store.dart';
@@ -33,6 +34,7 @@ void setupDI() {
   //Services
   getIt.registerSingleton<QubicHubService>(QubicHubService());
   getIt.registerSingleton<TimedController>(TimedController());
+  getIt.registerSingleton<BiometricService>(BiometricService());
 
   getIt.registerSingleton<QubicCmd>(QubicCmd());
 }
